Move FileLifecycleHooks down under Serilog.Sinks.File - avoids namespace pollution when using Serilog logging APIs; mark PeriodicFlushToDiskSink as obsolete in line with other public types shared with Serilog.Sinks.RollingFile
